karinacms/requirements.txt
Django==1.6
Markdown==2.3.1
PIL==1.1.7
South==0.8.4
argparse==1.2.1
distribute==0.6.24
djangorestframework==2.3.10
wsgiref==0.1.2
